Add cossorigin="anonymous" to main scss import
To avoid getting the error below, add the crossorigin attribute when you
access the site on http instead of https.

Error is:

The resource has an integrity attribute, but the resource requires the
request to be CORS enabled to check the integrity, and it is not. The
resource has been blocked.

Signed-off-by: Jason Rogena <jason@rogena.me>